Moved LED code to a seperate file. Now if you wanna use the LEDs, you just pass it into your module, like with the 'socket' code
